Problem: -------- After upgrade, InnoDB crashes on accessing FTS AUX Index tables. This is because in 5.7, FTS AUX Index table created the field "word" as either nullable or 'non-nullable' depending on the FTS index first field nullability.
Example:

5.7 AUX Index created with "word" field as "NULLABLE"
-----------------------------------------------------

CREATE TABLE t1 (a VARCHAR(200), b TEXT, FULLTEXT (a,b))
ENGINE = InnoDB charset utf8mb4;

Since the fts index first field is 'a' and it is nullable, FTS AUX index
tables are created with "word" field as nullable.

5.7 AUX Index created with "word" field as "NON NULLABLE"
-----------------------------------------------------

create table t1 (a varchar(200) not null, fulltext (a)) engine = innodb;

Since the fts index first field is 'a' and it is NON-nullable, FTS AUX index
tables are created with "word" field as NON-nullable.

In 8.0, the AUX Index tables are always 'NON-nullable'. So reading 5.7 FTS AUX
tables which are nullable, lead to crash.

Fix:
---
Similar to 5.7, create AUX Index tables based on nullablity of first field of
FTS index. This info is stored in column se_private_data.

Reviewed-By: Jimmy Yang <jimmy.yang@oracle.com>
RB: 17726