BUG#26576078: GR_PARALLEL_LOCAL_AND_REMOTE_TRANSACTIONS.INC CAN'T HANDLE SQL SYNC ERRORS
gr_parallel_local_and_remote_transactions.inc allow to test local transaction's
behavior (conflict/success) against a transaction that is ordered first on a
remote server and broadcast to local server.

The local_server_connection1 will execute a --send blocking the use of it.

A rpl_sync is execute on this process and if it fails will connect using a
server_* connection name returning error due not being able to execute a query
between a --send and a --reap.

To solve this issue we change the gr_parallel_local_and_remote_transactions.inc
to don't allow the use of server_* connection name on local_server_connection1.