BUG#26765648: MTR TEST FOR TESTING UNDO LOG TRUNCATION WITH 2 UNDO               TABLESPACES
Description:
------------
The current innodb_undo.truncate test causes size increase of only one
undo tablespace. The truncation of the undo tablespaces was being checked
by comparing strings that stored the sizes of the tablespaces before and
after server shutdown.

Fix:
---
Two transactions with sufficient DML are being used which causes size of
both tablespaces to increase. The truncation of the
tablespaces is being checked by using expressions that store and compare
the sizes of tablespaces before and after server shutdown.

Reviewed-by: Anitha Gopi <anitha.gopi@oracle.com>
Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>
Reviewed-by: Viswanatham Gudipati <viswanatham.gudipati@oracle.com>
RB:17607