Bug#27084922: ENABLE TESTCASES TO RUN DAILY ON PB2 WHICH ARE SKIPPING DUE TO NOT_LOG_BIN.INC
Many testcases are skipped when binary logging is ON. These were disabled as
part of WL#8794 (Qualify suites other than rpl and binlog with different
binlog modes and add to weekly PushBuild run), and few other instances.

With WL10470, the binary logging got enabled by default, and we changed
collection files to run the testcases with binary logging OFF weekly once.

Since, the actual issue of testcase failures in many cases was due to
different binlog_formats, we now plan to enable those testcases to run fine
with both log-bin and skip-log-bin. For code issues which are specific to
log-bin, the testcases will be enabled to run on daily pb2 by having skip-log-bin
in .opt files.

Reviewed by: sven.sandberg@oracle.com
Reviewed by: oystein.grovlen@oracle.com
Reviewed by: srikanth.b.r@oracle.com

RB: 17861