Bug #26588537 : INNODB.INNODB_BUFFER_POOL_RESIZE_DEBUG FAILS WITH SIG 6
Background:
Buffer pool resize is done in a separate thread. It is done in parallel with
other thread in which DMLs might be running on server. During resizing a buffer
pool following sequence happens:
- Disable AHI (if enabled)
- Resize buffer pool
- Enable AHI (if disabled earlier)

Analysis:
While disabling AHI, for all blocks in buffer pool, we try to reset index and
n_pointers=0 before we clear the AHI hash table.

BUT, during above processing, it is assumed that, all blocks in buffer pool
would be in BUF_BLOCK_FILE_PAGE state. Which is not true i.e. it is possible
that a block could be moved to a state BUF_BLOCK_REMOVE_HASH by some other
thread, while disabling AHI is in progress.

Fix:
During resizing of buffer pool, expect BUF_BLOCK_REMOVE_HASH state as well (with
BUF_BLOCK_FILE_PAGE) while disabling AHI.

RB: 17864
Reviewed by : satya.bodapati@oracle.com