BUG#27075816 - ASSERTION FAILURE: OS0FILE.CC:5127:!EXISTS
This is mainly due to table name was not parsed properly.
Table name stored in memory is in space name charset,
so there is a conversion from file name charset to space name
charset. The parser used to parse 't2#tmp' into 't2?tmp',
and this leads to the failure reported.

Solution is to fix this parser to always get correct 't2#tmp'.
So the trailing '#tmp' for non-partitioned table should be
considered in dd_parse_tbl_name().

Furthermore, two warnings/errors are suppressed for
ddl_crash_alter_partition, which can be ignored.

RB: 17907
Reviewed-by: Allen Lai <allen.lai@oracle.com>