BUG#27041552 - M_THD->GET_TRANSACTION()->IS_EMPTY(TRANSACTION_CTX::STMT)                ETC. ASSERTION
An event whose schedule happens with global autocommit OFF on expiry
results in an assert. The assert happens because on expiry the event is
dropped and as part of drop, the autocommit gurad constructor is called.
This guard asserts if there is an open pending transaction.
The drop of the event happens on expiry once the event schedule is executed.
This drop should happen in it's own transaction so the fix is to commit any
open transaction if any pending at time of drop.