Bug#25695434: MEMORY LEAK IN MULTI-SOURCE REPLICATION WHEN BINLOG_ROWS_QUERY_LOG_EVENTS=1
Problem:
--------

1. In multi source replication, Slave SQL thread leaks
memory when replicating using two channels from
master-master pair when binlog_rows_query_log_events is
enabled.

2. There is also a valgrind issue of invalid memory access
assosciated with this bug. i.e, show processlist on slave
accesses 'm_rows_query' even after it is deleted.

Analysis:
---------

For memory leak,
Though it may seem that it requires three
servers to reproduce the bug, but two servers are enough to
reproduce it.

1) When binlog_rows_query_log_event is enabled on master,
  Rows_query_log_event is created and written into binary
  log prior to Rows_log_event.

2) On slave, when reading from the relay log and applying
  it, the event itself is stored in
  Relay_log_info::rows_query_ev.

3) Usually Relay_log_info::rows_query_ev is deleted after
  execution of Rows_log_event by the function
  rows_event_stmt_cleanup() which is called in the end.

4) But whenever the statement must be skipped (because
  slave has already applied the event from another channel)
  it does not delete the rows_query_ev instance and leads to
  memory leak.

For valgrind issue,
After deleting the rows_query_ev and before resetting the
query, if someone gives 'SHOW PROCESSLIST', instead of
printing NULL, the original query is printed.

Fix:
----
For memory leak,
Replace 'DBUG_RETURN(0)' with a 'goto end' so that
rows_event_stmt_cleanup() function is called in the end
and the memory allocated for that instance is freed.

For valgrind issue,
The query is reset first and then the rows_query_ev is
freed.