Bug #22688065 ALTER TABLE EXECUTES COUNT(*) TO SEE IF ANY ROWS EXIST
Problem: ALTER TABLE does a call to handler::ha_records() in order to
find out if a table is empty or if it has rows. This call may cause a
table scan in InnoDB and is unnecessarily expensive.

Fix: Instead of retrieving the number of rows, try to read one row from
the table.

Change-Id: Ibdce319e1e408e165919ad1cc83be408f963a7f0