Bug #27101184: TEMPTABLE::TABLE::INDEXES_DESTROY CRASH DURING SHUTDOWN
During forced connection shutdown the temptables are not destroyed
before connection thread is finished. That starts the auto cleanup
of thread_local storage including tables collection in temptables.

The destructor calls a method to remove indexes which uses the
mysql_index. Value of that variable is a pointer stored from the
TABLE passed to temptable creation which is no longer present,
and the memory is no longer present at the moment of auto cleanup.

The fix removes usage of mysql_index and stores the required
index object allocation size in Table object.

Reviewed-By: Marcin Babij <marcin.babij@oracle.com>
Reviewed-By: Pawel Olchawa <pawel.olchawa@oracle.com>
RB: 17953