Bug#25965370 NDB : LQH CAN EXCEED 4 CONCURRENTLY   OPENED FILES PER PART DURING RESTARTS
During restart, the LQH component in the data node loads redo
log part metadata for each redo log part it manages, from one
or more redo log files.

Metadata is stored per megabyte, and each file has a limited
metadata capacity, so the number of files which must be consulted
depends on the size of the redo log part.

  (NoOfFragmentLogFiles * FragmentLogFileSize)

Files are opened, read and closed sequentially, but the close
of one file occurs concurrently with the open of the next.

In cases where file close is slow, this can result in more
than 4 open files per redo log part.

As the files are opened with the OM_WRITE_BUFFER option set,
this can result in more than 4 chunks of write buffer being
allocated per part.

As the write buffer pool is finite, if all redo log parts
are in a similar state, it may be exhausted, causing node
shutdown.

The solution chosen is to avoid using the OM_WRITE_BUFFER
option during metadata reload, so that transiently opening
more than 4 redo log files per part does not cause node
failure.

A new ERROR_INSERT is added to slow file closure for metadata
reload, allowing manual testing of the fix.  No automated
testcase is added as there is a requirement for large redo
logs.