BUG#25846957: MYSQL STARTUP BLOCKED FOR 20 MINUTES               ON PARTITION-ENGINE-CHECK.
Analysis:
=========
From MySQL-5.7.17, during the server start up,
a check is performed to list the tables which
uses non native paritioning. This helps users
to prepare for upgrade to 8.0 where the support
for non native partitioning is removed. For large
databases with many such tables, this check
consumed a lot of time since a SELECT on I_S is
performed to fetch the information. By default,
the variable 'disable-partition-engine-check' is
disabled, which triggers the check for tables
using non-native partitioning at startup.

Fix:
====
The variable 'disable-partition-engine-check' is enabled
by default. Hence the check for non native partitioned
tables is skipped during the start up unless it is
explicitly disabled.