Bug#27091493: GR_INJECT_PIPELINE_ERROR_LOCAL_GTID IS FAILING SPORADICALLY IN PB
Problem: Missing error supression when GR is unable to kill transaction.

The group replication plugin is stopping due to an error.
Sees that there is one waiting transactions but when it
tries to kill it, it fails, causing error being written to logs.
Thread is already killed or not in Wait_ticket map.

Resolution:
There is already suppressed error message in logs "Due to a plugin error, *".
Adding post intermittent message to suppression logs.