Bug #23590280   NO WARNING WHEN REDUCING INNODB_BUFFER_POOL_SIZE INSIZE 		THE FIRST CHUNK
PROBLEM

If we are scaling down innodb_buffer_pool_size inside the first chunk,
no resize operation is performed which is the correct behaviour . But
no warning is given to the client and no warning is included in the
error log . So to the user it looks as though the incorrect value has
been used to resize the buffer pool.

FIX

The above resize operation returns a warning to user.

Reviewed by: Jimmy and satya
RB: 17812