Bug#26762973: STORING SDI COSTS TOO MUCH TIME FOR A BIG PARTITIONED TABLE
Problem: When creating a partitioned table using implicit
tablespace, the effect is to place each partition in its own tablespace,
with no particular tablespace for the table as a whole. Since SDIs were
stored in all tablespaces used by a table, the cost of storing SDIs in a
table with many tablespaces became prohibitive.

Solution: Only include the tablespace for the first partition in the
set of tablespaces used to store the SDI.