Bug #27129545: WL#8478: NODE CRASHED DURING RESTART WITH ERROR "PGMAN (LINE: 2286) 0X00000002 C
The undo record data is stored in a buffer managed by LGMAN.
In DBTUP(0), a copy of this(through LinearSectionPtr) is sent to the LDM threads.
However, before this copy is made, there can be a signal sent to PGMAN from DBTUP(0).
This gives a chance to LGMAN to overwrite the undo record data buffer.
Hence, for cases where a signal is sent to PGMAN in DBTUP(0), save the buffer in c_proxy_undo.m_data
Also improve testability of the assumption that LGMAN assumes DBTUP does not rely on keeping
the buffer pointer valid.