Bug#26898143: RPL_ROW_JSONDIFF_ERROR FAILS SPORADICALLY ON WEEKLY-8.0 AND WEEKLY-TRUNK
 Test is failing because there is an unsupressed
ERROR logged in the error log of slave server. The
test has supression for ERRORS 4550 and 4551 which
are supposed to be caused due to the scenarios in
test. Recent change in the server by some commit
has changed the ERROR numbers for different errors.
So the error which was causing ERROR no 4550, 4551
now is causing ERROR no. 4549 and 4550 respectively.

Fix: Included the supression for Error no 4549.
     Removed the supression for Error no 4551.