Bug#24444908: CLUSTER CRASHED DURING RESTART WITH AN ASSERTION !CALLBACKOBJ->HAS_DATA_TO_SEND
and

Bug#25128512 SET RR WITH >1 NG FAILS DUE TO 'WE(1) HAVE BEEN DECLARED DEAD BY 2 (VIA 4)'

Patch fixes race conditions related to asynchronous disconnect / reconnect
of the transporter while there could be other threads concurrently
inserting signal data into the send buffers.

A ::reset_send_buffer() was performed as part of a disconnect, and
we required the send buffers to still be empty when we later reconnect.
When this failed, we could either hit the '!has_data_to_send' assert,
or we received malicous data which had survived the disconnect.

This patch introduce a number of changes to the Transporter implementations
in order to fix the insufficent concurrency controll:

1)
 The virtual TransporterCallback method reset_send_buffer() is
replaced with the methods enable_send_buffer() / disable_send_buffer().
As part of this an 'enabled' state is also introduced in each implementation
of the TransporterCallback. This enable us to discard any thread local
send buffers which are 'late' in being appended to a disabled transporter.

1a)
 Enable and disabling of send buffers has been moved into the
Transporter::connect* methods instead of having them in the
TransporterRegistry::report_connect/_disconnect. The *receivers*
are inited here as their connection state is protected by the
poll right. However, the poll right gives no such protection
for the send buffers. (Added a large comment about this)

Thus, it feels more correct to handle
the send buffers where the transporters are connected.

1b)
 The method has_data_to_send() was removed as part of 1). As any
'data_to_send' could later be discarded by disabling a node, it
didn't really make sense to check for 'has_data' is it could
immediately go away before doSend or whatever. Similar logic has
instead been added to get_bytes_to_send_iovec(), where we with necessary
locks held can collect (or discard) send buffer to doSend().

2) Added the TransporterSendBufferHandle interface method ::isSendEnabled().
It replace the usage if Transporter::isConnected() where this method was
used in a 'send context'. isSendEnabled() provides us with a thread safe
way of checking if send is enabled for a specific 'send handle'.

2a) Usage of Transporter::isConnected in TransporterRegistry::prepareSend()
is replaced with ::isSendEnabled().

2b) Defined a 'synchronized' (client threads) or 'optimistic' (data nodes)
implementation of ::isSendEnabled(). The 'synchronized' implementation
should be the 'safe' implemetation which Transporter::isConnected() never
was. This provides us with a thread safe way of checking the connectivity
to a specific node. (API clients)

2c) Datanodes (both mt- and non-mt) never really cared about a SEND_DISCONNECTED
error from prepareSend() as it was ignored in the SimulatedBlock::sendSignal()
methods. (Handled as SEND_OK together with SEND_BLOCKED). So an 'optimistic'
::isSendEnabled() is implemented here. (A 'synchronized' implememtation
with have required taking a mutex, which we probably cant afford)
Any send buffers allocated data to disabled nodes are discarded by the
frequent doSend() calling get_bytes_to_send_iovec()