Bug#26422857 GCS CAN USE SOME VERBOSITY IN THE ERROR LOG
PROBLEM
-------
As part of the fix of Bug#25495393:

(a) We identify several low-level log messages; and
(b) We identify desired, unexistant, log messages.

We should remove or modify the messages in (a) because they are unhelpful to
users. We should include the messages in (b) because they are helpful.

ANALYSIS
--------
We performed the majority of the analysis as part of the fix of Bug#25495393.

FIX
---
Only log the unhelpful messages in (a) when GCS is built for debug.
Specifically, swap the usage of G_MESSAGE for G_DEBUG.
Modify the message informing that XCom is ready to accept connections.
Log the messages in (b). Specifically, the highest known XCom protocol
version.

Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed-by: Filipe Campos <filipe.campos@oracle.com>
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
RB: 18038