BUG#26962434 -- XA INCONSISTENCIES AFTER XA COMMIT ERROR
This is a temporary workaround for the bug#26962434 to make the test
binlog.binlog_xa_commit_failure pass ok.

General idea underlying this workaround is to add server restart at the end of
the test case, rolls back prepared XA transaction if it exists after restart
and suppress possible warnings about presence of prepared XA transaction
in server's log.