Bug #27128524: ASSERTION `!TABLE || (!TABLE->READ_SET                .. FIELD_INDEX))' AT FIELD.CC:5438
Issue:
------
1) This problem occurs when a table's column is present in
   the read set of a trigger.
2) skip-log-bin option is used.
3) LOAD statement is used to execute the trigger.

As part of the fix for Bug #27041382,
Table_trigger_dispatcher::mark_fields() was removed from
prepare_triggers_for_insert_stmt(). This was handled for
INSERT statement by adding it to the prepare_inner.

The problem for LOAD was masked because when binlog is open
all the columns are marked as being part of the read_set.
So when the skip-log-bin option is used, all the columns
aren't marked as being part of the read_set. It hits an
assert during val_str for a column that isn't part of the
read_set.

Solution:
---------
Add the call to Table_trigger_dispatcher::mark_fields() for
LOAD.