Bug#26568667: TEST INNODB.UNDO_ENCRYPT_BOOTSTRAP IS UNSTABLE
Issue:
------
innodb.undo_encrypt_bootstrap fails sporadically with a warning that keyring_file
plugin is busy and will be uninstalled on shutdown.

Fix:
-----
The test restarts the server at the end, ensuring that the plugin is uninstalled
and proper clean up is done. The warning is supressed since we know for sure that
the plugin will be uninstalled upon complete execution of the test.