BUG#27166613 Ensure ATRT waits for processes to be undefined before (re)start
While CPCD is of asynchronous nature (i.e. when requesting to start a process
does not necessarily mean the process has started when the method returns)
ATRT assumed CPCD was syncrhonous. Hence, when ATRT restarted a process it
could happent that it would start a process before it has completely stop
leading to test run failures.

The main change was to add logic to wait for a process to stop. This logic
must be used after stopping a process or a set of processes.
We did not make wait for process stop part of the stop logic to prevent
refreshing state every time we want to stop a process and to allow stopping
several processes in one go, and then waiting for those to finish also in
one go.

The above change additionally required that process status update is done
via update_status method only, hence emoving wrong assumptions about the
processes life-cycle (e.g. when saving a process, first perform stop, and
then save its status).

Additionally, we made CpcClient connection more robust by checking that
messages are actually sent and received (prevent erroneous usage of the
API before connection is open).

Finally, added extra logging to ATRT to be able to more easily diagnose test
failures.

NOTE: There's an interesting situation when resetting processes configuration,
      we start by stopping the management nodes first and then the data nodes.
      Stopping the second last data node, causes the last data node to quit
      by itself (watchdog). We updated the logic not to fail on this scenario.
      However, a better approach would be to reverse all this logic, stopping
      first all the servers, then the data nodes, and last the mgmd nodes.