wl#11164 Rework first patch in this series.
The first patch introduced: 'Always return a TRANSID_AI with correlation-id to scan queries'.
This was done with the intention that INNER join needed thge correlation-is to
determine which parent row we didn't find any matches for.

As this is only needed for an INNER-joined TreeNode, we do not need
this extra TRANSID_AI returned for a non-INNER join.