Bug#27067082:RPL.RPL_MTS_LOGICAL_CLOCK_APPLIER_ERROR_CONCURRENT_STOP_SLAVE - PB2 FAILURE
Problem:
A call to an include that issues "STOP SLAVE" may happen when the slave thread
are already stopped by errors in slow machines (or testing with valgrind).
In such cases, the test case will produce an extra content not recorded
into its result:

+Warnings:
+Note   3084    Replication thread(s) for channel '' are already stopped.

Fix:
The test case is not 100% deterministic, so to avoid the above warning we
can wrap the include that issues "STOP SLAVE" with
disable_warnings/enable_warnings.