Bug#27168722: Speed up the GET_JSON_WEIGHT function in the json_conversions test
Make GET_JSON_WEIGHT a simple CASE statement to avoid a table scan for
each call. The json_conversions test relies heavily on the
GET_JSON_WEIGHT function, and this change reduces the time it takes to
run the test from 100 seconds to 30 seconds with a debug build in my
environment.

Change-Id: I660f4c0ea41050b5c47953285e843a95dacbc797