Bug#27121663: SIG 11 IN FILL_DD_VIEW_TABLES | SQL/DD/DD_VIEW.CC
The name of a derived table was not saved before the table was
materialized and assigned the name of the temporary table. When the
table name was later to be restored to its original value, an invalid
reference (nullptr) was set. This caused the server to crash because
'strlen' was called with nullptr as argument.

As a fix, add a call to save_name_temporary() before materialization
such that the existing call to reset_name_temporary() after
materialization resets the name correctly.

Also do some cleanup related to internal table names. Previously, "*"
was used as an internal table name for derived tables, but it was later
decided to just use the empty string "". Some unnecessary code was still
leftover after this change. The internal_table_name buffer is therefore
removed, and length of table names for views and derived tables should
be 0 instead of 1.