Bug#27136346 SYSBENCH CONNECT TEST SHOWS -95% REGRESSION FOR NON-ROOT VS ROOT USER
None root users experienced a 95% drop in CPU utilization due to a cache
which is suppose to store database privileges based on user name and host
for fast acccess. The cache is protected be a rw lock that needs to be
checked for success. The check was inveresed so it reported failure
on success and vice versa which ultimately caused the performance issues
as a write lock was taken 10 times for every connection attempt by
internal system processes.