Bug#26576922: SERVER DOESN'T ALWAYS FALL BACK TO DEFAULT ERROR MESSAGES
Server was changed so it would start even if lc-messages-dir
contained an invalid path; it would then fall back on the
built-in error messages. The intention was primarily to have
the server come up even in the absence of lc-messages-dir
(i.e. to require as few start-up arguments as possible, ideally
just the data-dir); that the server would throw an error, but
no longer abort on invalid values was just a welcome side effect.

The same was not true for lc-messages (the actual language),
as that had a sensible default value (and therefore satisfied
the above requirement, "server works with argument omitted").
Paul noted that users might find it confusing that providing
an invalid value throws a non-fatal error for one case
(lc-messages-dir), and a fatal one for the other (lc-messages).

Therefore, we now throw an error but continue for all three of
lc-messages-dir, lc-messages, and lc-time-names.