Bug#27169809 FIX BUG 17583 FOR ALPINE LINUX
This is an extra addition to the bugfix described below.
To ensure consistent behaviour also on Alpine, open an extra file descriptor.

Also: when reporting errors in batch mode, flush stdout before flushing stderr.
This makes the output consistent for commands like
-- exec mysql test>&1 < "<path to input file>"

Similarly for mysqldump: flush the result file before dumping any error
reports, to make output deterministic.

Bug#17583: mysql drops connection when stdout is not writable

When the client program had its stdout file descriptor closed by the calling
shell, after some amount of work (enough to fill a socket buffer) the server
would complain about a packet error and then disconnect the client.

This is a serious problem.  If stdout is closed before the mysql is
exec()d, then the first socket() call allocates file number 1 to communicate
with the server.  Subsequent write()s to that file number (as when printing
results that come back from the database) go back to the server instead in
the command channel.  So, one should be able to craft data which, upon being
selected back from the server to the client, and injected into the command
stream become valid MySQL protocol to do something nasty when sent /back/ to
the server.

The solution is to close explicitly the file descriptor that we *printf() to,
so that the libc layer and the OS layer both agree that the file is closed.

Change-Id: Ic74a2cb752d1a66ff6d7e8e9b8fb98fb85636051