Bug#27001934 - DEMERGE CODE SNIPPETS FROM SQL_TABLE, WHICH ARE NEEDED BY MEB
Server code, needed by MEB was cut out of server files
by #ifndef UNIV_HOTBACKUP.

This patch reduces the amount of InnoDB code, compiled with
MEB, so that the functions, which need server code, are not
included. The changes in server code are reverted.

Unrelated changes:

Fix Windows build: Define strcasecmp and strncasecmp as
inline functions instead of preprocessor macros.

Change-Id: Iaabf9800f9e5c72282c312f76ad8da2765663c3f