Bug #27175716: KEYRING MIGRATION TOOL GENERATING THE KEYS WHICH CAN'T                BE READ BY SERVER
Problem: Keyring migration tool creates backed file for destination
plugin if the file is not present. This file is created with OS user
as owner for this file. After migration when server is started with
destination keyring plugin, plugin will reject this file from reading.

Fix: Moved the migration specific code after calls to setgid and setuid
so that backend files for destination plugin is created with correct user.