Bug#27121500 PERFORMANCE_SCHEMA INDEX ON OBJECT_TYPE BROKEN
Queries in performance schema using indexes on OBJECT_TYPE
columns do not return the correct results.

The root cause is the evaluation of whether a given record
matches or not a given key value,
implemented in PFS_key_object_type::match().
This code does not implement NULL checks properly.

In the performance schema, a record that has a NULL
OBJECT_TYPE is represented with an object_type == NO_OBJECT_TYPE,
and there are no boolean flags to mark the field as NULL.
This works because NO_OBJECT_TYPE is not a possible value.

In a key provided by a WHERE clause,
the object type comes from the query, and can have any value,
especially for VARCHAR columns.

When a key is NULL, PFS_key_object_type::m_is_null is true,
and the value of PFS_key_object_type::m_object_type is irrelevant.

When a key is not null, but when the OBJECT_TYPE provided
(as in WHERE OBJECT_TYPE = 'impossible')
does not match the expected object types,
the code converts the string to NO_OBJECT_TYPE,
exposing the bug.

The bug is that this comparison:
  m_object_type == object_type
does not work when a key is not null but impossible,
as it matches records that are not supposed to match.

The fix is to implement PFS_key_object_type::do_match(),
with proper checks for NULL keys and or NULL records
before comparing the key value with the record value.

Likewise for PFS_key_object_type_enum.