Bug#21616914: MAIN.JSON_CONVERSIONS TIMES OUT IN VALGRIND
Strip down the time it takes to run the json_conversions test by
making the following changes:

1. Remove some redundant verification of the contents in the test
tables. The contents of a big test table were dumped, and multiple
regular expressions were used on the output to filter out
non-deterministic data. The input data were validated in different and
cheaper ways later in the test, so this partly redundant and very time
consuming step was removed.

2. A table that was populated by doing multiple INSERT INTO ... SELECT
DISTINCT FROM statement that used subqueries to avoid inserting
duplicates into the table, is now instead populated using a single
INSERT INTO ... SELECT statement where the SELECT is a big UNION. This
single statement is much faster than the original ones.

3. Add an index on a column that is used for equality predicates in
many of the queries in the test.

4. Testing of the JSON comparator with the != operator is removed. The
non-standard != operator is indistinguishable from the standard <>
operator after the query has gone through the tokenizer, so it doesn't
contribute to the code coverage.

5. Some parts of the test, which use multiple SELECT statements to
work through a table, are rewritten to use a single SELECT statement
which goes through the entire table in a single pass.

6. Remove some duplicated SELECT statements that tested conversions
from the YEAR data type.

Change-Id: I55bfbea5fad26cf0feba717ef32b43c4e88c1457